EditTreePageの子にTreeAreaとToolAreaが来るようにコンポーネント切り出し・コンポーネント名を変更
